input: atmel_tsadcc: enable touchscreen averaging and add fast wake up
Enable the touchscreen average to improve the resulting events. For this
to work the trigger period needs to be reduced.

This puts a field into at91_tsadcc_data to allow platforms to specify
the number of conversions to average over.

XXX: should the trigger period passed by the platform, too, as this
depends on the number of conversions?
XXX: seperate fast wake up into a seperate patch? What does it?
XXX: don't use bare constants

Signed-off-by: Ludovic Desroches <ludovic.desroches@atmel.com>