net: bus: Add implementation of Bus domain sockets
Each process connected to an AF_BUS socket, has one or more addresses
within that bus. These addresses are 64-bit unsigned integers,
interpreted by splitting the address into two parts: the most significant
16 bits are a prefix identifying the type of address, and the remaining
48 bits are the actual client address within that prefix.

The process that calls bind(2) on the socket is the owner of the bus and
is called the bus master. The master is a special client of the bus and has
some responsibility for the bus' operation. The master is assigned a fixed
address with all the bits zero.

The prefix with all bits zero is reserved for use by the kernel, which
automatically assigns one address from this prefix to each client on
connection. The address in this prefix with all bits zero is always
assigned to the bus master. Addresses on the prefix 0000 are unique
and will never repeat for the lifetime of the bus master.

AF_BUS transport can support two network topologies. When a process
first connect to the bus master, it can only communicate with the
bus master. The process can't send and receive packets from other
peers on the bus. So, from the client process point of view the
network topology is point-to-point.

The bus master can allow the connected peer to be part of the bus
and start to communicate with other peers by setting an socket option
with the setsockopt(2) system call using the accepted socket descriptor.
At this point, the topology becomes a bus to the client process.

Packets whose destination address is not assigned to any client are routed
by default to the bus master (the client accepted socket descriptor).

Signed-off-by: Javier Martinez Canillas <javier.martinez@collabora.co.uk>